+++ Project/Dtos/MealProductDto.cs
@@ -4,6 +4,6 @@
     {
         public int MealId { get; set; }
         public int ProductId { get; set; }
-        public float QuantityInGrams { get; set; }
+        public double QuantityInGrams { get; set; }
     }
 }
