+++ Project/Models/MealProduct.cs
@@ -4,7 +4,7 @@
     {
         public int MealId { get; set; }
         public int ProductId { get; set; }
-        public float QuantityInGrams { get; set; }
+        public double QuantityInGrams { get; set; }
 
         public Meal Meal { get; set; }
         public Product Product { get; set; }
