+++ Project/Models/Product.cs
@@ -4,10 +4,10 @@
     {
         public int ProductId { get; set; }
         public string Name { get; set; }
-        public float Calories { get; set; }
-        public float Protein { get; set; }
-        public float Fat { get; set; }
-        public float Carbohydrates { get; set; }
+        public double Calories { get; set; }
+        public double Protein { get; set; }
+        public double Fat { get; set; }
+        public double Carbohydrates { get; set; }
 
         public ICollection<MealProduct> MealProducts { get; set; }
     }
